Create Cube and adjust cube 4. Cube 4 can not work will fix it later

+++ Assets/Game/Scripts/RubikCube/CubePieceStateData.cs
@@ -3,9 +3,9 @@ using System;
 namespace Game.Scripts.RubikCube {
     [Serializable]
     public class CubePieceStateData {
-        public int coordinatorX;
-        public int coordinatorY;
-        public int coordinatorZ;
+        public float coordinatorX;
+        public float coordinatorY;
+        public float coordinatorZ;
 
         public float positionX;
         public float positionY;
